VTX-3411: some tests
